+++ moPepGen/svgraph/SubgraphTree.py
@@ -11,13 +11,15 @@ if TYPE_CHECKING:
 class SubgraphBranch():
     """ ParentGraphLocation """
     def __init__(self, _id:str, level:int, parent_id:str,  location:FeatureLocation,
-            variant:VariantRecord):
+            variant:VariantRecord, feature_type:str, feature_id:str):
         """ constructur """
         self.id = _id
         self.level = level
         self.parent_id = parent_id
         self.location = location
         self.variant = variant
+        self.feature_type = feature_type
+        self.feature_id= feature_id
 
     def is_parent(self, other:SubgraphBranch) -> bool:
         """ Checks if it is the parent of a given subgraph """
@@ -36,23 +38,26 @@ class SubgraphTree():
         """ Get item """
         return self.data[key]
 
-    def add_root(self, _id:str):
+    def add_root(self, _id:str, feature_type:str, feature_id:str, variant:VariantRecord):
         """ Add the main graph """
         subgraph = SubgraphBranch(
-            _id=_id, level=0, parent_id=None, location=None, variant=None
+            _id=_id, level=0, parent_id=None, location=None, variant=variant,
+            feature_type=feature_type, feature_id=feature_id
         )
         self.data[_id] = subgraph
         self.root = subgraph
 
     def add_subgraph(self, child_id:str, parent_id:str, level:int,
-            start:int, end:int, variant:VariantRecord):
+            start:int, end:int, variant:VariantRecord,
+            feature_type:str, feature_id:str):
         """ Add subgraph """
         location = FeatureLocation(
             seqname=parent_id, start=start, end=end
         )
         subgraph = SubgraphBranch(
             _id=child_id, level=level, parent_id=parent_id,
-            location=location, variant=variant
+            location=location, variant=variant,
+            feature_type=feature_type, feature_id=feature_id
         )
         self.data[child_id] = subgraph