+++ moPepGen/svgraph/__init__.py
@@ -6,3 +6,4 @@ from moPepGen.svgraph.TVGCursor import TVGCursor
 from moPepGen.svgraph.PeptideVariantGraph import PeptideVariantGraph
 from moPepGen.svgraph.PVGNode import PVGNode
 from moPepGen.svgraph.ThreeFrameCVG import ThreeFrameCVG
+from moPepGen.svgraph.VariantPeptideTable import VariantPeptideTable